[Refactor] 댓글 컴포넌트 리팩토링, 게시글 상세 페이지 게시글 영역 댓글 영역 분리
